Refactor - move classes to Util
