+++ src/main/java/edu/harvard/iq/dataverse/util/PersonOrOrgUtil.java
@@ -9,9 +9,6 @@ import javax.json.JsonObject;
 import javax.json.JsonObjectBuilder;
 import javax.json.JsonString;
 
-import edu.harvard.iq.dataverse.export.openaire.Cleanup;
-import edu.harvard.iq.dataverse.export.openaire.FirstNames;
-import edu.harvard.iq.dataverse.export.openaire.Organizations;
 import edu.harvard.iq.dataverse.util.json.JsonUtil;
 import edu.harvard.iq.dataverse.util.json.NullSafeJsonBuilder;
 
@@ -69,7 +66,7 @@ public class PersonOrOrgUtil {
      * @return
      */
     public static JsonObject getPersonOrOrganization(String name, boolean organizationIfTied, boolean isPerson) {
-        name = Cleanup.normalize(name);
+        name = StringUtil.normalize(name);
 
         String givenName = null;
         String familyName = null;
